server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
scipy>=1.13.0
pydantic>=2.7.0
python-multipart>=0.0.9
Pillow>=10.3.0
matplotlib>=3.8.0
